Split out again, minor clean up
